fix: node install command to use correct version
When we added the new `sn_interface` crate, we forgot to update these functions for selecting the
version number of sn_node. This would have been caught by our test for this command, but
unfortunately that test can't run because it's subject to random failures due to rate limiting from
the Github API.

The README for the release process is also updated to include this as a step.