chore: move request_operation_fulfilled
This function is moved from the `operations` module to the top level module, since the `operations`
and other modules that now only contain tests, will be getting removed. The tests in the modules
being removed will be moved into the `detection` module.

Some unit test coverage was added for this function, and a new `get_unfulfilled_ops` function was
added to facilitate easier testing, but it could also be used by callers of the API. It made testing
easier because it wraps the code that reads values from the concurrent data structures, which can be
quite verbose.