ci: use different mechanism for crate publishing
The last two release runs have failed when attempting to publish `sn_node`. It seems even if the
dependent crate is returned in a `cargo search`, this doesn't ensure the crate will still be
available when you attempt to publish the crate with the dependency.

Now the script is changed to simply just attempt the publish itself in a retry loop.