chore: simplify cleanupPeerLinks
There was a suspected deadlock in the CleanUpPeerLinks code, so here
we simplify things in order to hopefully prevent any deadlock.

Moving the cleanup into comms, removing any checks against membership
(as all nodes should be connectable; clients can always retry).

And removing PeerLinks that are not conncted at all.