fix: publishing script not exiting correctly
In the last release run, publishing of crates were successful with the retry loop; however, the
script didn't exit correctly on the successful publish. The loop then continued until the retries
were exceeded.

The problem was the use of the sub shell. I use this from habit whenever a script is changing
directories, as the change only applies inside the sub shell, meaning the outer script will retain
its current directory, which is often desirable behaviour. The use of `exit 0` inside the sub shell
only exits the sub shell, not the entire script, as was intended.

I've now removed the use of the sub shell; it wasn't really a necessary protection for this scenario
anyway.