feat: command to generate secret key for dbc owner
A new `keys create-dbc-owner` command is added to generate a secret key for use with reissuing DBCs.
The key gets serialised to a file in `~/.safe/cli`. It will be read from here when making
connections to the network.

This commit also features various other changes. Originally I had two separate commits, but I
accidentally amended a previous commit and couldn't separate the content.

The `keys create` command is using functions from a module related to authentication to determine
where to write the keys it generates. These functions directly specified a credential file's
location at the user's home directory. To help unit testing, I refactored this to provide the
location from the `Config` struct. Previously I refactored `networks` commands in a similar fashion.
The result is, during the unit test, the `assert_fs` library will create a temp directory for the
config and so the user's real home directory won't be manipulated during the test suite. `assert_fs`
will also remove the temp directory at the end of each test.

With this change in place, I provided a couple of unit tests for the `keys create` command. This
command doesn't hit the network, so integration tests aren't necessary. There are only really two
code paths to go through: one for writing the keypair to file and one for not doing so.

I didn't test the `keys show` command because it mostly just prints information to the console,
and at this stage I don't know how to capture that for a unit test (or if it's even possible).

Finally, I tidied up some commented out code that doesn't seem to apply any more, and added some
additional documentation for the commands to try and make their intent a bit clearer.