feat: use persistent dbc owner in sn_api
The network connection functions in the CLI are updated to pass the persistent DBC Owner. This is
based on the existence of a key that was created with the new `create-dbc-owner` command. If the key
exists, the connection will use the persistent DBC owner; if `None` is passed, a random key will be
generated.

During a reissue, the change DBC is then assigned the owner from the client. We don't yet assign the
output DBC this owner, because we're still only supporting bearer DBCs. This may change for owned
DBCs, which should come soon.

I split the wallet reissue test out into a couple: one with a single input DBC and one with
multiple. The original test had multiple inputs, but it seemed it was worth checking that the
reissue would function if there were also just a single input.