feat: add bls type to keypair enum
Extends the `sn_interface::types::keys::Keypair` enum to support the BLS keypair type. We need this
because we want to change the CLI to use BLS rather than Ed25519 keys, so we can support signing
output DBCs with the same keypair we use to sign commands sent from the CLI.

I modified the tests that check each keypair type can be serialized/deserialized. Previously there
was one test case looping over each type of keypair, but I think it's better style to define each
test case explicitly: you are calling out what scenarios you want to support and it makes the cases
easier to understand at a glance, even if there is a small bit of repetition between them.