feat!: reissue dbc to a particular owner
BREAKING CHANGE: the `wallet_reissue` API is extended to add an `owner_public_key` argument. This is
an `Option<bls::PublicKey>`. To reissue to a bearer DBC, use `None`, but to reissue to a particular
owner, specify the key.

I've also taken an opportunity to update the API documentation a little. In the doc comments and
logging output, I changed "Wallet" references to "wallet": in this context, "wallet" is not a proper
noun, so it shouldn't be capitalised (unless it happens to be at the start of a sentence). If you
wanted to refer to the type `Wallet`, that would be capitalised, but it should also be referred to
using enclosing backticks.