feat!: wallet deposit read input dbc from stdin
BREAKING CHANGE: the positional `dbc` argument is changed to an optional argument.

The DBC hex representation is 8000+ characters long, so this gives the user the option to provide
the large input DBC from a file, like so:
```
safe wallet deposit <safe-url> < ~/.safe/input_dbc
```

Bash seems to allow you to provide this huge argument on the command line, but it doesn't work in
Fish, which is the shell I use.