feat: add public key argument for owned dbcs
The `wallet reissue` command now has an additional optional argument, `--public-key`, which allows
the user to reissue a DBC to be owned by the holder of that public key. The key should be BLS
hex-encoded.

The `wallet deposit` command will now require extension to provide the secret key when depositing an
owned DBC. This will be done as a separate piece of work.

Some additional changes were made in support or to tidy CLI-related code:
* The conversion of DBCs to/from hex were removed from the CLI since this is now done on the `Dbc`
  type.
* A CLI test that existed to test the above conversion code was removed since it's no longer
  necessary.
* The naming scheme for the CLI wallet tests were elaborated and the redundant "calling_safe"
  prefixes were removed.