feat: dbc arg can refer to file or dbc data
The `--dbc` argument for the `wallet deposit` command is extended so it can refer to either a file
path or DBC data. It will be tested first for an existing file path, otherwise it will fall back to
being interpreted as DBC data.

Some people have been having trouble with pasting DBC data directly via the command line, depending
on what OS or shell is in use. You can still also pass the data via stdin, but that's only available
on Linux or macOS.