ci: prevent sn_cli being published incorrectly
In the last release run, the release failed because it attempted to publish sn_cli, which did not
have its version bumped.

The reason was because the test for doing a publish is to check the commit message contains the
crate name, and it just so happens that "sn_cli" is contained in "sn_client". In the case of this
release, sn_client had changes, but sn_cli didn't, so the attempted publish of sn_cli failed.