feat!: extend wallet_deposit for owned dbcs
BREAKING CHANGE: the `wallet_deposit` function now provides an `Option<bls::SecretKey>` argument for
providing a secret key when depositing an owned DBC.

The owned DBC is changed to a bearer by providing the secret key, and the resulting DBC is then
stored in the wallet. This means you don't need to provide the secret key at reissue time, where
there is an issue with mapping input DBCs to secret keys.

A couple of misc changes:
* Extra test cases were added here to cover assigning the name to the deposit.
* More instances of "Wallet" were converted to "wallet", because "wallet" is not a proper noun.

As an unrelated change, this commit also provides more test coverage for `keys show` command. This
isn't a hugely important command, but I just wanted to ensure we don't break support for BLS keys.

These are integration tests because you really need to parse the console output, which you can't do
with unit tests.

This should probably have been its own commit, but it was accidentally added via an amend, then the
API changes were also amended on top, so it was too difficult to break out.