feat!: serialize to bls keys in util functions
Utility functions were recently added to the API for serializing to the `Keypair` type. This was
changed to serialize directly to BLS to make it easier for the CLI to deal directly with BLS keys.
Soon we will be refactoring the `Keypair` type to have a different use case and things like
`sn_client` would be refactored to directly work with BLS keys. This is a little step in that
direction.

There was a utility function added to `sn_interface` to create a `Keypair` from a hex-based BLS key
because we still need to use the `Keypair` at this point in time.