feat!: remove use of xorurl with keys command
The use of BLS keys means the XorUrl that was displayed to the user will no longer be usable, since
the BLS key is a 48-byte structure, and not 32 bytes like the Dalek key. We will come back and
address this issue later, possibly re-introducing the `SafeKey` type in some kind of different form.

The functionality and corresponding test cases for the `cat` and `dog` command that were working
with the `SafeKey` data type were removed.

The test cases for the `keys` commands were elaborated here to check the output of both the pretty
print and the json formats and to make sure the CLI prints out a matching keypair.