feat: extend cli wallet deposit for owned dbcs
The CLI is now extended to support the deposit of owned DBCs.

The `deposit` command will check if the supplied DBC is owned, and if it is, it will check to see if
the `--secret-key` argument is present and use that. If that argument isn't present, it will attempt
to use the secret key that's configured for use with the CLI, i.e., the `keys create --for-cli`
command.

The `reissue` command was also extended to provide an `--owned` flag, which when used, will reissue
an owned DBC using the public key configured for use with the CLI. This argument is mutually
exclusive with the `--public-key` argument, which will reissue the DBC using a specified key.

So we could offer the user a suggestion when a supplied secret key didn't match, this also involved
making a little extension to the API, to return a specific type of error. We will need to modify
`sn_dbc` to return a specific error type for this too, so we can avoid checking the string content
of the error message, but this will be covered on a separate PR.