fix(rejoin): disable rejoins
It seems that we may have not fully thought through the implications
of nodes rejoining.

The flow was:
1. a node leaves the section (reason is not tracked)
2. the node rejoins
3. the section accepts them back and attempts to relocate them

This seems fine until you realize that relocating the node may fail
for some reason and now we're stuck holding this re-joined node.

Also we don't track why this node was removed from the section.

They could have been a faulty elder that now is back in the same
section it had attacked with the same age which will likely cause it
to become an elder again.