chore(node): simplify the send funcitonality
Previously we've had a send to client and a send ot nodes, the latter
which would only send to a subset (with the delivery group setting), and
reported errors.

We've stripped out the delivery group stuff already as unnecessary. So here
we just have one function to send from nodes, any errors are handled
appropriately if the recipient is in our section, otherwise ignored.