ci: use system-wide location for genesis dbc
Previously the file was being placed under the profile directory, but the test process was
complaining it couldn't find the file. The custom action for running Nextest should be running on
the same machine and user, meaning it really shouldn't be a problem to put the file there, but
nonetheless it apparently couldn't be found. Thus, I'm going to try a system-wide location.

Also changed the mechanism for setting the variable to be defined on the invocation of the Nextest
action.