chore(dysfunction): remove unused severity; refactor weighted score
Prev weighted score related everything to the std_deviation, but this
has the effect of nullifying outliers and decreasing the impact of
weighting.

Instead we opt for a simple "threshold" score, above which, we're
dysfunctional. So the sum of all issues tracked is used, and if
we reach above this point, our node is deemed dysfunctional.