refactor: clean up unused functionality
`closest` is a method that will find a prefix that is closest, but if
not returning any, it means the set is empty. The `closest_or_opposite`
used this function internally, but actually never got to the opposite,
because `closest` would always return a SAP.

This method was used in a few places where no exclusions were given, so
it is clear in that case that it would always find a prefix. In a single
case, it was called with an exclusion, where it would find a section
closer than its own section.