ci(nightly): slack notifications for nightly run
When the run fails, use a custom action to post a message to a Slack incoming webhook. For some
reason the content of the message to be sent is defined using environment variables, so you are
limited in terms of the amount of information that can be included in the notification. I checked a
few more actions and they all operated in this fashion. With that in mind, the best thing I thought
to do was for the failure to have a direct link to the run, so we can get people to the logs as
quickly as possible.

Also made a little change to output the direct URL of the server logs that get uploaded.