feat: return error to client on unknown section key
If one of the spent proofs sent by the client have been signed with a key this section is not
currently aware of, return an error back to the client.

This introduces a new SpentProofUnknownSectionKey variant to the messaging data errors, because none
of the existing variants seemed appropriate for this scenario.