chore: switch on clippy::unwrap_used as a warning
sn_interface:
All usages were in testing code. They were either removed in favour of using the `?` operator or an
`ok_or_else` with an error message. The `allow` attribute was applied for a couple of cases where
the usage was in a `proptest!` macro. This was agreed to be preferable to using `except`.

sn_dysfunction:
All usages were in `proptest!` macros and therefore the `allow` attribute was applied.

sn_node:
A conversion from `sn_dbc::Error` to `sn_node::storage::Error` was added to remove an unwrap in
the bootstrapping process. The remainder were in test cases, with one `allow` added for use in a
`proptest!` macro.

sn_client:
A couple of usages were found in little pattern matching functions. They were changed to return an
IO error, which is the error type in the `Result`, the return type of the closure the code is called
in. It wasn't possible to use the `eyre!` macro here because there was no conversion between the IO
error and the `eyre` error type.

sn_api:
One usage was found while calculating the total amount for DBC reissues. The arithmetic for the
`Token` type returns an `Option` with `None` in the case of an overflow. The unwrap was replaced
with an error.

sn_cli:
Removed several usages from testing code. All usages were the unwrapping of an `Option`, which was
replaced with `ok_or_else` and an error message.