feat: use err result to send errors back to client
Previously the mechanism for handling errors in `handle_valid_service_msg` was to return an `Ok`
result with an empty list of commands, or if you wanted to send an error response back to the
client, to include a command in that `Ok` result.

A new `CmdProcessingClientRespondError` variant is added to `Error` to enable returning an `Err`
result when we want to return an error back to the client. When the command processing code
encounters this error variant, it will process the commands included in the error, which will be the
information to send back to the client.

The intention is also to return `Err` results from `handle_valid_service_msg` when any errors are
encountered, not just for the limited few we want to send back to the client.