refactor: err result from gen spent proof share
The spent proof share generation function now returns `SpentProofShare` directly rather than an
`Option`, and we return `Err` results rather than `Ok(None)`. The unit tests were then updated to
check for specific kinds of errors. The calling code is also a bit cleaner since it can just use the
`?` operator.

Note the use of an `allow` attribute to suppress a Clippy warning about this function having too
many arguments. The additional arguments are just required for the composition of the error that
gets sent back to the client, which I consider justified. Otherwise this error will need to be
handled in a special way outside the function, which is an unnecessary complication.