chore: makefile targets for easier local network use
There are now two targets for running a local network, one that does a clean build and one that
doesn't. The nodes have debug logging set and also use a debug build for speed.

Also added some additional logging output to the command processing code to indicate the return of
an error to the client.

This also fixes up a mistake made while resolving a merge conflict.