feat: dbc spend can update network knowledge
When a DBC is submitted for spending, there are a set of spent proofs sent along with the request,
one for each input for the DBC being spent. It's possible that these proofs were signed with a
section key that the section processing the request is not aware of. In this case, the request
processing section needs to be updated with new network knowledge. Therefore, we add a couple of
fields to the spend command to give it the option to send new network knowledge, namely the updated
proof chain and the signed section authority provider.

If the node detects the spend command has been sent with these, it will create a new type of
internal command, `UpdateNetworkAndHandleValidServiceMsg` which will be returned for processing.
This command will contain the original spend request, which will be run after the network knowledge
is updated. The reason for issuing the new command is because the original `HandleValidServiceMsg`
command does not have mutable access to the node.