feat: retry dbc spend on unknown section key
This is the client side for the scenario where the spent proofs are signed by section keys that the
processing section is not aware of.

Several retries will be attempted because it's possible for there to be multiple section keys that
are not known, but the network will only return back one key at a time.

Based on review feedback, this commit also changes the `SpentbookCmd::Spend` variant to use a single
field for the updated network knowledge, along with some other more minor changes.