tests: spend with updated network knowledge
Previously I had a placeholder in for this case, but now have something working.

The test requires having two network sections and one of the input DBCs for a transaction being
signed by the other section key.

The `TestNodeBuilder` was extended with a function that creates a section without a creating a node,
and this included being able to provide a section chain and tree.