chore: remove spend retry on client
The spend retry depends on providing new network knowledge. We will be using another mechanism to
obtain this knowledge, which is not available at the moment. Once it's available, we'll add the
retry again.

For now we decided it's best to remove it and only merge the node-side changes.

This also fixes up various changes after the merge of the new SectionsDAG that replaced the
SecuredLinkedList.