refactor: get public commitments from sn_dbc
The code for retrieving and validating the public commitments was moved out to the sn_dbc crate.

It's needed for both the spend request and test setup code which is going to be referenced in both
`sn_node` and `sn_client`.

Also fixed a clippy error in `SectionTree::get_signed_by_key`.