feat: client retry spend on unknown section key
When the client receives an unknown section key error, it will obtain the proof chain and SAP for
the unknown section and resubmit the request with this additional information.

There is no automated client test for this scenario. We went to great lengths to try, but it proved
not worth the effort. It was too difficult to modify the network knowledge with a fake section and
still have things function correctly. The scenario is unit tested on the node side, and we done
enough testing to know that the retry loop does what's intended.

There are a few misc changes along with this commit:

* Debugging message to indicate a spend request being processed correctly, which proved useful when
  trying to get the automated test working.
* Remove the current section key from the unknown section key error. It's not necessary to include
  this.
* When running the baby fleming network with the Makefile, include log messages from `sn_interface`.
* Fix up git-based references to `sn_dbc` crate.