refactor: rename authority field to owner
This was something we discussed in a team meeting.

Since a new `SignedRegister` type will be introduced which will have an `authority` field with a
section signature, it was more appropriate for the current `authority` field to be renamed to `owner`.

The `SignedChunk` type also uses `authority`, so that keeps it consistent with that.