feat: sign registers with section key
The register creation command already had a `section_sig` field on it, but this was being populated
with a dummy value. It has now been changed to an `Option` field. It is assigned `None` when the
register creation command is created on the client, then when it's received by the node, the field
is populated with the section key.

BREAKING CHANGE: in similar fashion to the `SignedChunk` type, a new `SignedRegister` type is
introduced. Register storage now uses this new type and queries also return it.