ci: use s3_deploy user for nightly and release workflows
In the nightly workflow, this user gets used to launch the testnet (really just reading the tfstate
file) and to upload node logs to the `sn-node` bucket when a run fails.

In the release workflow it's used to upload the built artifacts to the same `sn-node` bucket.

The names of the variables were changed to make it more specific about which user we were referring
to.

The keys being used previously have been deleted.