ci: use single thread for cli/api tests
There are issues on Windows with the `fs::rename` function when the test suite is running with
multiple threads.

Code in the CLI related to the config file is calling this API to rename a temporary file to
`config.json` and on Windows you get an access denied error when there are multiple threads doing
this in quick succession.

The real solution here is probably to isolate the config directory for each test, because I don't
think it's a realistic user scenario for multiple instances of the CLI to be running at the same
time on the same machine in such close time proximity that a user would actually run into issues
here.

There are also issues with ACK messages not being returned correctly that may have something to do
with running the test suite with multiple threads, so the API tests are also set to use a single
thread.