chore(node): use STANDARD_CHANNEL_SIZE in node. increase to 100_000
Channels could be an uniwitting blocker as we await for capacity.

If we expect channels to be cleared in a timely fashion,
we should aim not to have an artifical bottleneck.

100_000 chosen as vast majority of channelsdon't require a lot of data.
And anything large waiting, will just be in a channel elsewhere taking up memory,
so this way we avoid slowing down node processing