feat: provide `--config-dir-path` arg for `safe`
Add a global `--config-dir-path` argument that will set the location of the config directory. If
it's not set, the `SAFE_CONFIG_DIR_PATH` environment variable will be used, and if that's not set,
we will default to `$HOME/.safe`.

The purpose of this customisation is for integration testing so that each test can use an isolated
configuration directory, but users can apply it too if they wish.