test: use isolated config dir for cli tests
A function is added to the test utils library for creating a temporary directory that functions as
the config directory for the CLI. It uses `assert_fs`, so the directory will be automatically
deleted when the reference to it is dropped. This enables each test case to use its own config
directory, which resolves issues we've had with multiple instances of `safe` writing/reading the
config file in a race situation.

The functions for running `safe` commands were changed to require the config directory to be
passed. This forces each test case to use this mechanism.

All the tests were updated to use the same naming convention, with the exception of the `files get`
tests, which are quite tedious and probably best left with the convention they were using, at least
for now.

Some tests could be un-ignored as part of this change.

The API and CLI test suites had their single-thread limitation removed.