ci: only version bump on nightly workflow
Removes the test suite from the nightly workflow and instead only performs a version bump.

Shortly we will change this workflow to occur on a push to main, but to make sure the release
process is functioning correctly it will be easier initially if this only occurs once a day.