test(fault): simplify failure check, make msgs target only one node
Previously we had messages mapping to several nodes depending on the issue type...
But actually, the msg counts are for _all_ nodes, so this was unuecessarily
levelling the field regardless of node quality.

Now one message goes to one node, and it's one failure test.
Which means we see more realisitic variance in scores and the stats will
not be quite as arbitrary