ci: add arm builds to merge workflow
We seen a failure on a release because an ARM build failed. The failure occurred because the `cross`
utility (which we use for cross compilation) wouldn't build correctly. Strangely, I wasn't able to
reproduce this problem, but I think it highlights the need for building on ARM before we merge. If
we are going to release on every merge, we will need to check that we can build on ARM.