fix: correctly trigger failure notifications
The `failure` function in Github Actions only checks the failure of previous *steps* in a job, so it
can only really be used on a per job basis.

It's cumbersome to have this on every job, but there appears to be no alternative way. Luckily there
aren't many release-related jobs in the relevant workflows.