refactor: idle_timeout from 10s to 70s
This was the timeout before this pull request. 70s was deduced to be a
value that gave CI the time to pass many tests. Although this value is
not optimal in a real world scenario, as routers might close the holes
where the connections rely on, and thus the connections will actually
timeout sooner in some cases. An optimal value was previously decided to
be 18s, as some routers supposedly close holes after 20s.