refactor: leave out reachability check for join
This reachability check tries to reply/reach out to the node that could
not connect, but we only save incoming connections, so this will not
make it out to the node.

Also, the reachability check was left out of the last release of qp2p,
as the network should not only rely on these reachability checks, but
the nodes should be unable to join the network anyway if they're not
reachable in the first place.