refactor(node): pass public addr in --first
The genesis node will output a network contact file which contains its
own address. This public address was specified by a separate flag but
was removed, making the genesis incapable of producing a proper contacts
file. This changes the `--first` flag to be used to pass in that public
address, which should be done for all genesis nodes, so the proper
external address can be advertised.