chore: introduce sn_updater library crate
This library crate provides utility code to be shared between the `safe` and `sn_node` binaries so
they can update themselves to the latest version.

We can no longer use the `self_update` crate to do this with a single statement because
`self_update` is designed for use with single-crate repositories, though we do still make use of
utility code it happens to make public.

The crate makes an `update_binary` function available. The doc comments in the diff provide details
of how it works, so I won't repeat that here.