fix: restore sn_node update command
The updating mechanisms for the `safe` and `sn_node` binaries stopped working when we switched to
using a workspace repository.

We now make use of the newly introduced `sn_updater` crate to provide a working solution.

The start up behaviour of the node changes slightly here. The code that validated the arguments was
moved to occur after the updating process because that validation doesn't apply when you're just
performing a simple update, which will terminate the node process after it's done. In addition, the
`--update-only` argument was removed, because you always need to restart the node after it has been
updated. I didn't see much purpose in this argument and also though it could potentially cause
confusion.