fix: restore safe update command
The `safe` binary also makes use of the newly introduced `sn_updater` crate to provide a working
update mechanism.

Important note: there is code in the `node update` command that could be replaced with `sn_updater`,
but the whole `node` command is due to be removed anyway, which is what I will be working on next.

Also fixes up a few things after a rebase from main.