feat: reworked version of testnet binary
The `testnet` binary was reworked to remove its dependency on `sn_launch_tool`. Since it will now be
used as a replacement for the `node run-baby-fleming` command, the interface was updated to make it
a bit more usable. Now it does not attempt to build the node by default: the `--build-node` flag is
used for that. The other arguments that were added should be fairly self explanatory and can be seen
on the diff.

All the support from the original binary is retained in this rework, including the use of flamegraph
and joining an existing network.

There is a module in the crate that exposes a couple of structs that can be used for launching a
testnet: `TestnetBuilder` and `Testnet`. The functions on `Testnet` should have almost full code
coverage, but since the builder is a simple utility, I didn't bother putting tedious tests around
it. Since the bulk of the work has been unit tested I also haven't added any integration tests for
testing the code in the `main` module for the binary, since it's pretty simple and doesn't really
seem worth the additional testing infrastructure.

The merge CI workflow was updated to use the reworked version of the binary. Since we already build
the node in the previous step, we just pass the path to it, rather than attempt to rebuild.

The script for verifying node joins was modified to check for IPs matching "0.0.0.0" in addition to
"127.0.0.1", which was being caused by the use of the `--local-addr` argument. It also now considers
the total number of nodes to be the genesis node plus the number of non-genesis nodes. In addition,
the script was updated to split long commands over multiple lines, use newer-style Bash syntax and
consistent indentation size.