chore: remove the `node` command from `safe`
BREAKING CHANGE: the `node` command is entirely removed.

As agreed previously between both the dev team and community users, the `node` wrapper command is
entirely removed in favour of using the `safenode` binary directly.

All the associated code and helper code was removed. This functionality will now be provided by the
`safeup` and `testnet` tools.