chore: rename testnet crate to sn_testnet
Even though the `testnet` crate name is not taken on crates.io, I think it makes sense to prefix
this crate with `sn_`, as per our other crates. The name of the binary does not change. This crate
needs to be published because `sn_client` has a dependency on it.

This also provides a README for the crate, which was necessary to have it published.