ci: rename sn_cli asset name
Renames the `sn_cli` assets on the Github Release from `sn_cli` to `safe`. This was supposed to be
done as part of the last modification to the process, but I seem to have accidentally missed it.

I'm also re-enabling the release and version bumping workflows to run on push now that I've cleared
out the last bad release.